+++ openstack/dcs/v1/lifecycle/Get.go
@@ -4,6 +4,6 @@ import "github.com/opentelekomcloud/gophertelekomcloud"
 
 // Get a instance with detailed information by id
 func Get(client *golangsdk.ServiceClient, id string) (r GetResult) {
-	_, r.Err = client.Get(client.ServiceURL("instances", id), &r.Body, nil)
+	raw, err := client.Get(client.ServiceURL("instances", id), nil, nil)
 	return
 }
