+++ openstack/dcs/v1/others/GetProducts.go
@@ -9,6 +9,6 @@ import (
 // Get products
 func GetProducts(client *golangsdk.ServiceClient) (r GetResult) {
 	// remove projectid from endpoint
-	_, r.Err = client.Get(strings.Replace(client.ServiceURL("products"), "/"+client.ProjectID, "", -1), &r.Body, nil)
+	raw, err := client.Get(strings.Replace(client.ServiceURL("products"), "/"+client.ProjectID, "", -1), nil, nil)
 	return
 }
