+++ openstack/dcs/v1/others/ListAvailableZones.go
@@ -9,6 +9,6 @@ import (
 // Get available zones
 func ListAvailableZones(client *golangsdk.ServiceClient) (r GetResult1) {
 	// remove projectid from endpoint
-	_, r.Err = client.Get(strings.Replace(client.ServiceURL("availableZones"), "/"+client.ProjectID, "", -1), &r.Body, nil)
+	raw, err := client.Get(strings.Replace(client.ServiceURL("availableZones"), "/"+client.ProjectID, "", -1), nil, nil)
 	return
 }
