+++ openstack/dcs/v1/others/ListMaintenanceWindows.go
@@ -9,6 +9,6 @@ import (
 // Get maintain windows
 func ListMaintenanceWindows(client *golangsdk.ServiceClient) (r GetResult3) {
 	// remove projectid from endpoint
-	_, r.Err = client.Get(strings.Replace(client.ServiceURL("instances/maintain-windows"), "/"+client.ProjectID, "", -1), &r.Body, nil)
+	raw, err := client.Get(strings.Replace(client.ServiceURL("instances/maintain-windows"), "/"+client.ProjectID, "", -1), nil, nil)
 	return
 }
